third_party/WebKit/LayoutTests/webaudio/OfflineAudioContext/offlineaudiocontext-suspend-resume-sequence-expected.txt

Test OfflineAudioContext.resume() and OfflineAudioContext.suspend() with the timed sequence.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Scheduling suspend #0 at 0.25 second(s).
PASS Scheduling suspend #1 at 0.75 second(s).
PASS Scheduling suspend #2 at 1 second(s).
PASS Scheduling suspend #3 at 0.5 second(s).
PASS Scheduling suspend #4 at 1.25 second(s).
PASS Scheduling suspend #5 at 0 second(s).
PASS Scheduling suspend #6 at 1.75 second(s).
PASS The resolution order of suspend #5 is 0 at 0.00 second(s).
PASS The resolution order of suspend #0 is 1 at 0.25 second(s).
PASS The resolution order of suspend #3 is 2 at 0.50 second(s).
PASS The resolution order of suspend #1 is 3 at 0.75 second(s).
PASS The resolution order of suspend #2 is 4 at 1.00 second(s).
PASS The resolution order of suspend #4 is 5 at 1.25 second(s).
PASS The resolution order of suspend #6 is 6 at 1.75 second(s).
PASS successfullyParsed is true

TEST COMPLETE
